refactor(yape06): de-duplicate yape06 module (#47)
Seventh de-duplication step of #47, following the established pattern.

- src/yape06/yape06.ts: replace the type-only stub with the REAL
  implementation moved verbatim from the monolith (detect + laplacian /
  min-eigen-value thresholds). Imports compute_laplacian and
  hessian_min_eigen_value from ./yape06_utils (already a real module).
- src/jsfeatNext.ts: shrinks by ~85 lines.

Verified behavior-preserving: tsc --noEmit clean; npm test 57/57 (the yape06
parity test pins detect() against original jsfeat); UMD bundle smoke-checked
(instanceof, default thresholds, detection on a synthetic image).
